:bug: 유틸 정리 및 불필요한 코드 정리
