+++ src/main/java/kr/co/antoon/common/util/CommonUtil.java
@@ -1,0 +1,5 @@
+package kr.co.antoon.common.util;
+
+public class CommonUtil {
+    public final static String APPLICATION_JSON_UTF_8 = "application/json;charset=UTF-8";
+}
