+++ src/main/java/kr/co/antoon/common/util/TimeUtil.java
@@ -1,0 +1,10 @@
+package kr.co.antoon.common.util;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class TimeUtil {
+    public static String now(String format) {
+        return LocalDateTime.now().format(DateTimeFormatter.ofPattern(format));
+    }
+}
