+++ src/main/java/kr/co/antoon/graph/presentation/GraphScoreController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import static kr.co.antoon.common.Utility.APPLICATION_JSON_UTF_8;
+import static kr.co.antoon.common.util.CommonUtil.APPLICATION_JSON_UTF_8;
 
 @Api(tags = "그래프 API")
 @RestController
